Split unit-tests into their respective configuration types
